huffmanCode.cpp, array.h and priorityQueue.h fix

+++ LibraryCPP/array.h
@@ -5,7 +5,7 @@
 
 // Non-resizeable array
 // Stores integer values inside
-typedef int Data;
+typedef unsigned long long int Data;
 
 struct Array;
